+++ instrumentation/opentelemetry_bandit/lib/opentelemetry_bandit.ex
@@ -34,7 +34,7 @@ defmodule OpentelemetryBandit do
                       default: [],
                       type_spec: quote(do: opt_in_attrs()),
                       doc: """
-                      Use semantic conventions library to ensure compatability, e.g. `[HTTPAttributes.http_request_body_size()]`
+                      Use semantic conventions library to ensure compatibility, e.g. `[HTTPAttributes.http_request_body_size()]`
 
                       #{Enum.map_join(opt_ins, "\n\n", &"  * `#{inspect(&1)}`")}
                       """
@@ -93,7 +93,7 @@ defmodule OpentelemetryBandit do
                     ]
                   )
 
-  @typedoc "Use semantic conventions library to ensure compatability, e.g. `HTTPAttributes.http_request_body_size()`"
+  @typedoc "Use semantic conventions library to ensure compatibility, e.g. `HTTPAttributes.http_request_body_size()`"
   @type opt_in_attr() ::
           unquote(ClientAttributes.client_port())
           | unquote(HTTPAttributes.http_request_body_size())
@@ -119,7 +119,7 @@ defmodule OpentelemetryBandit do
   ### Opt-in Semantic Convention Attributes
 
   Otel SemConv requires users to explicitly opt in for any attribute with a
-  requirement level of `opt-in`. To ensure compatability, always use the
+  requirement level of `opt-in`. To ensure compatibility, always use the
   SemConv attribute.
 
   Example:
@@ -314,12 +314,11 @@ defmodule OpentelemetryBandit do
   end
 
   defp set_network_protocol_attrs(attrs, conn) do
-    case extract_network_protocol(conn.adapter) do
-      {:error, _reason} ->
-        attrs
-
-      {:http, version} ->
-        Map.put(attrs, NetworkAttributes.network_protocol_version(), version)
+    case Plug.Conn.get_http_protocol(conn) do
+      :"HTTP/1.0" -> Map.put(attrs, NetworkAttributes.network_protocol_version(), :"1.0")
+      :"HTTP/1.1" -> Map.put(attrs, NetworkAttributes.network_protocol_version(), :"1.1")
+      :"HTTP/2" -> Map.put(attrs, NetworkAttributes.network_protocol_version(), :"2")
+      _ -> attrs
     end
   end
 
@@ -348,22 +347,6 @@ defmodule OpentelemetryBandit do
         )
     end
   end
-
-  defp extract_network_protocol(
-         {_adapter_name, %{transport: %{__struct__: Bandit.HTTP1.Socket} = transport}}
-       ) do
-    case transport.version do
-      :"HTTP/1.0" -> {:http, :"1.0"}
-      :"HTTP/1.1" -> {:http, :"1.1"}
-      nil -> {:error, "Invalid protocol"}
-    end
-  end
-
-  defp extract_network_protocol({_adapter_name, %{transport: %{__struct__: Bandit.HTTP2.Stream}}}) do
-    {:http, :"2"}
-  end
-
-  defp extract_network_protocol(_), do: {:error, "Invalid protocol"}
 
   defp parse_method(method) do
     case method do
@@ -411,25 +394,12 @@ defmodule OpentelemetryBandit do
 
     case otel_http_extract_client_info(client_headers, config[:client_headers_sort_fn]) do
       %{ip: :undefined} ->
-        %{adapter: {Bandit.Adapter, adapter}} = conn
-
-        extract_client_info_from_transport(adapter.transport)
+        %{address: peer_address, port: peer_port} = Plug.Conn.get_peer_data(conn)
+        %{ip: ip_to_string(peer_address), port: peer_port}
 
       client_address ->
         client_address
     end
-  end
-
-  defp extract_client_info_from_transport(%{__struct__: Bandit.HTTP1.Socket} = transport) do
-    start_meta = transport.socket.span.start_metadata
-
-    %{ip: ip_to_string(start_meta.remote_address), port: start_meta.remote_port}
-  end
-
-  defp extract_client_info_from_transport(%{__struct__: Bandit.HTTP2.Stream} = transport) do
-    {ip, port} = transport.transport_info.peername
-
-    %{ip: ip_to_string(ip), port: port}
   end
 
   defp otel_http_extract_client_info(headers, nil) do
@@ -498,15 +468,35 @@ defmodule OpentelemetryBandit do
       |> Tracer.set_attributes()
     end
 
+    maybe_update_span_name(conn)
+
     Tracer.end_span()
     Ctx.clear()
+  end
+
+  defp maybe_update_span_name(conn) do
+    case conn.private do
+      %{plug_route: {route, _fun}} ->
+        method = sanitize_method(conn.method)
+        Tracer.update_name("#{method} #{route}")
+        Tracer.set_attribute(HTTPAttributes.http_route(), route)
+
+      _ ->
+        :ok
+    end
+  end
+
+  defp sanitize_method(method) do
+    if parse_method(method) == HTTPAttributes.http_request_method_values().other,
+      do: "HTTP",
+      else: method
   end
 
   @doc false
   def handle_request_exception(meta, config) do
     Tracer.set_status(OpenTelemetry.status(:error, ""))
 
-    Tracer.record_exception(meta.exception, meta.stacktrace)
+    record_exception(meta)
 
     # bandit does not set this on the meta but extracts this after the exception
     # telemetry is emitted
@@ -514,12 +504,29 @@ defmodule OpentelemetryBandit do
 
     %{
       HTTPAttributes.http_response_status_code() => status_code,
-      ErrorAttributes.error_type() => meta.exception.__struct__
+      ErrorAttributes.error_type() => error_type(meta.exception)
     }
     |> set_resp_header_attrs(meta.conn, config)
     |> Tracer.set_attributes()
 
     Tracer.end_span()
     OpenTelemetry.Ctx.clear()
+  end
+
+  defp record_exception(%{exception: exception, stacktrace: stacktrace})
+       when is_exception(exception) do
+    Tracer.record_exception(exception, stacktrace)
+  end
+
+  defp record_exception(%{kind: kind, exception: reason, stacktrace: stacktrace}) do
+    :otel_span.record_exception(:otel_tracer.current_span_ctx(), kind, reason, stacktrace, [])
+  end
+
+  defp error_type(%struct_name{} = reason) when is_exception(reason) do
+    struct_name
+  end
+
+  defp error_type(reason) do
+    reason
   end
 end